packages/aserto/pyproject.toml
[tool.poetry]
name = "aserto"
version = "0.21.1"
description = "Aserto API client"
readme = "README.md"
authors = ["Aserto, Inc. <pypi@aserto.com>"]
maintainers = ["authereal <authereal@aserto.com>"]
homepage = "https://github.com/aserto-dev/aserto-python/tree/HEAD/packages/aserto"
repository = "https://github.com/aserto-dev/aserto-python/tree/HEAD/packages/aserto"
documentation = "https://github.com/aserto-dev/aserto-python/tree/HEAD/packages/aserto"
license = "Apache-2.0"
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]
packages = [
    {include = "aserto", from = "src"}
]

[tool.poetry.dependencies]
aiohttp = "^3.8.0"
python = "^3.8"
typing-extensions = "^4.8.0"
grpcio = "^1.49.1"
protobuf = "^4.21.7"
aserto-authorizer = {path = "../../../python-authorizer", develop = true}
aserto-directory = {path = "../../../python-directory", develop = true}

[tool.poetry.dev-dependencies]
black = "21.7b0"
isort= "^5.9.3"
mypy = "^1.6.1"
pytest-asyncio = "^0.15.1"
pyright = "^1.1.336"
requests = "^2.31.0"

[tool.black]
line-length = 100
target-version = ["py37"]

[tool.isort]
profile = "black"

[tool.mypy]
show_error_codes = true
check_untyped_defs = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_expr = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
implicit_reexport = false
no_implicit_optional = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "test.mock"
disallow_any_expr = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

